FA-206 Handled mismatch between component.multiple and data
If a component is switched from single -> multiple, then that data is wrapped in an array after it is loaded.

If a component is switched from multiple -> single, it will default to the first value of the old array data ONLY when viewing/editing the submission. It will still show the old array data in the table view until it is saved as a single value. (new directive is added to components that support multiple to do this conversion)