FA-487 Replaced jquery.maskedinput with angular-ui-mask
We are currently using a forked version of angular-ui-mask that has one PR merged to fix an issue where clicking on an inputmask field without typing anything would make it $dirty and trigger validation. That looks like it'll get merged soon so whenever it does, i'll point us back to the original repo.